requirements.txt
pytest
pytest-cov
typer
pythonocc-core=7.8.1.1
pyqt
